+++ src/main/java/tconstruct/mechworks/entity/item/EntityLandmineFirework.java
@@ -86,7 +86,7 @@ public class EntityLandmineFirework extends Entity {
         if (this.prevRotationPitch == 0.0F && this.prevRotationYaw == 0.0F) {
             float f = MathHelper.sqrt_double(par1 * par1 + par5 * par5);
             this.prevRotationYaw = this.rotationYaw = (float) (Math.atan2(par1, par5) * 180.0D / Math.PI);
-            this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, (double) f) * 180.0D / Math.PI);
+            this.prevRotationPitch = this.rotationPitch = (float) (Math.atan2(par3, f) * 180.0D / Math.PI);
         }
     }
 
@@ -110,7 +110,7 @@ public class EntityLandmineFirework extends Entity {
         float f = MathHelper.sqrt_double(this.motionX * this.motionX + this.motionZ * this.motionZ);
         this.rotationYaw = (float) (Math.atan2(this.motionX, this.motionZ) * 180.0D / Math.PI);
 
-        for (this.rotationPitch = (float) (Math.atan2(this.motionY, (double) f) * 180.0D / Math.PI);
+        for (this.rotationPitch = (float) (Math.atan2(this.motionY, f) * 180.0D / Math.PI);
                 this.rotationPitch - this.prevRotationPitch < -180.0F;
                 this.prevRotationPitch -= 360.0F) {
         }
